pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0.0"
addopts = "-vv -s --tb=native"
testpaths = "tests/"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "| %(asctime)s | %(name)s | %(levelname)s | %(filename)s | %(message)s"

[tool.mypy]
plugins = "returns.contrib.mypy.returns_plugin"

# We cannot work without explicit `Any` types and plain generics:
# disallow_any_explicit = True
# disallow_any_generics = True

allow_redefinition = false
check_untyped_defs = true
disallow_untyped_calls = true
disallow_subclassing_any = true
follow_imports = "silent"
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
local_partial_types = true
warn_no_return = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true



[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["zhukovgreen <iam@zhukovgreen.pro>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
pydantic = "^1.8.2"
attrs = "^21.2.0"
cattrs = "^1.8.0"
returns = "^0.16.0"
mypy = "^0.910"
types-PyYAML = "^5.4.10"
types-setuptools = "^57.4.0"
types-toml = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
